change `scarb completions` to regular command
