--- extensions/scarb-completions/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "scarb-completions"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-clap = { workspace = true, features = ["derive"] }
-clap_complete.workspace = true
-anyhow.workspace = true
-scarb = { path = "../../scarb" }
-scarb-ui = { path = "../../utils/scarb-ui" }
-scarb-cli = { path = "../../utils/scarb-cli" }
-
-[dev-dependencies]
-scarb-test-support = { path = "../../utils/scarb-test-support" }
-indoc.workspace = true
